Deprecate non-indy plugins
