+++ apm-agent-core/src/main/java/co/elastic/apm/agent/bci/HelperClassManager.java
@@ -32,8 +32,6 @@ import net.bytebuddy.dynamic.ClassFileLocator;
 import net.bytebuddy.dynamic.loading.ByteArrayClassLoader;
 import net.bytebuddy.dynamic.loading.ClassInjector;
 import net.bytebuddy.dynamic.loading.PackageDefinitionStrategy;
-import net.bytebuddy.matcher.ElementMatcher;
-import net.bytebuddy.pool.TypePool;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -43,10 +41,8 @@ import java.lang.ref.WeakReference;
 import java.security.ProtectionDomain;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.WeakHashMap;
@@ -80,8 +76,12 @@ import java.util.WeakHashMap;
  * Note: trying to load the helper class implementations from the bootstrap classloader leads to {@link NoClassDefFoundError}s.
  * </p>
  *
+ * @deprecated Use {@link TracerAwareInstrumentation#indyPlugin() indy plugins} instead.
+ * As the whole package the indy plugin is defined it is loaded from {@link IndyPluginClassLoader},
+ * a separate helper class manager is not needed anymore.
  * @param <T> the type of the helper interface
  */
+@Deprecated
 @VisibleForAdvice
 public abstract class HelperClassManager<T> {
     private static final Logger logger = LoggerFactory.getLogger(HelperClassManager.class);
@@ -146,8 +146,12 @@ public abstract class HelperClassManager<T> {
      * which has to be collectible when the web application is un-deployed.
      * </p>
      *
+     * @deprecated Use {@link TracerAwareInstrumentation#indyPlugin() indy plugins} instead.
+     * As the whole package the indy plugin is defined it is loaded from {@link IndyPluginClassLoader},
+     * a separate helper class manager is not needed anymore.
      * @param <T> the type of the helper class interface
      */
+    @Deprecated
     public static class ForSingleClassLoader<T> extends HelperClassManager<T> {
 
         @Nullable
@@ -203,8 +207,12 @@ public abstract class HelperClassManager<T> {
      * In order to optimize performance, stale entries are removed only when new helpers are being loaded (normally at the beginning of
      * startup time and when new applications are deployed). These maps shouldn't grow big as they have an entry per class loader.
      *
+     * @deprecated Use {@link TracerAwareInstrumentation#indyPlugin() indy plugins} instead.
+     * As the whole package the indy plugin is defined it is loaded from {@link IndyPluginClassLoader},
+     * a separate helper class manager is not needed anymore.
      * @param <T>
      */
+    @Deprecated
     public static class ForAnyClassLoader<T> extends HelperClassManager<T> {
 
         // doesn't need to be concurrent - invoked only from a synchronized context
@@ -275,60 +283,6 @@ public abstract class HelperClassManager<T> {
             }
             return helper;
         }
-    }
-
-    public static class ForIndyPlugin {
-
-        private static final Map<ClassLoader, Map<Collection<String>, WeakReference<ClassLoader>>> alreadyInjected = new WeakHashMap<ClassLoader, Map<Collection<String>, WeakReference<ClassLoader>>>();
-
-        /**
-         * Creates an isolated CL that has two parents: the target class loader and the agent CL.
-         * The agent class loader is currently the bootstrap CL but in the future it will be an isolated CL that is a child of the bootstrap CL.
-         */
-        public synchronized static ClassLoader getOrCreatePluginClassLoader(@Nullable ClassLoader targetClassLoader, List<String> classesToInject, ClassLoader parent, ElementMatcher<? super TypeDescription> exclusionMatcher) throws Exception {
-            classesToInject = new ArrayList<>(classesToInject);
-
-            Map<Collection<String>, WeakReference<ClassLoader>> injectedClasses = getOrCreateInjectedClasses(targetClassLoader);
-            if (injectedClasses.containsKey(classesToInject)) {
-                ClassLoader pluginClassLoader = injectedClasses.get(classesToInject).get();
-                if (pluginClassLoader == null) {
-                    injectedClasses.remove(classesToInject);
-                } else {
-                    return pluginClassLoader;
-                }
-            }
-
-            List<String> classesToInjectCopy = new ArrayList<>(classesToInject.size());
-            TypePool pool = new TypePool.Default.WithLazyResolution(TypePool.CacheProvider.NoOp.INSTANCE, ClassFileLocator.ForClassLoader.of(parent), TypePool.Default.ReaderMode.FAST);
-            for (Iterator<String> iterator = classesToInject.iterator(); iterator.hasNext(); ) {
-                String className = iterator.next();
-                if (!exclusionMatcher.matches(pool.describe(className).resolve())) {
-                    classesToInjectCopy.add(className);
-                }
-            }
-            logger.debug("Creating plugin class loader for {} containing {}", targetClassLoader, classesToInjectCopy);
-
-            Map<String, byte[]> typeDefinitions = getTypeDefinitions(classesToInjectCopy, parent);
-            // child first semantics are important here as the plugin CL contains classes that are also present in the agent CL
-            ClassLoader pluginClassLoader = new IndyPluginClassLoader(targetClassLoader, parent, typeDefinitions);
-            injectedClasses.put(classesToInject, new WeakReference<>(pluginClassLoader));
-
-            return pluginClassLoader;
-        }
-
-        private static Map<Collection<String>, WeakReference<ClassLoader>> getOrCreateInjectedClasses(@Nullable ClassLoader targetClassLoader) {
-            Map<Collection<String>, WeakReference<ClassLoader>> injectedClasses = alreadyInjected.get(targetClassLoader);
-            if (injectedClasses == null) {
-                injectedClasses = new HashMap<>();
-                alreadyInjected.put(targetClassLoader, injectedClasses);
-            }
-            return injectedClasses;
-        }
-
-        public synchronized static void clear() {
-            alreadyInjected.clear();
-        }
-
     }
 
     static Class injectClass(@Nullable ClassLoader targetClassLoader, @Nullable ProtectionDomain pd, String className, boolean isBootstrapClass) throws IOException, ClassNotFoundException {
