+++ apm-agent-core/src/main/java/co/elastic/apm/agent/bci/VisibleForAdvice.java
@@ -11,9 +11,9 @@
  * the Apache License, Version 2.0 (the "License"); you may
  * not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *   http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
@@ -39,7 +39,10 @@ import java.lang.annotation.Target;
  * the signature of the advice method,
  * as well as the advice class itself need to be public.
  * </p>
+ * @deprecated This annotation is not needed for {@link TracerAwareInstrumentation#indyPlugin()}  indy plugins}.
+ * That's because indy plugins never use {@link Advice.OnMethodEnter#inline() inline}d advices.
  */
+@Deprecated
 @Retention(RetentionPolicy.SOURCE)
 @Target({ElementType.TYPE, ElementType.FIELD, ElementType.METHOD, ElementType.CONSTRUCTOR})
 public @interface VisibleForAdvice {
