+++ apm-agent-core/src/test/java/co/elastic/apm/agent/bci/InstrumentationTest.java
@@ -524,6 +524,11 @@ class InstrumentationTest {
         public Collection<String> getInstrumentationGroupNames() {
             return List.of("test", "experimental");
         }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
+        }
     }
 
     public static class MathInstrumentation extends TracerAwareInstrumentation {
@@ -547,6 +552,11 @@ class InstrumentationTest {
         public Collection<String> getInstrumentationGroupNames() {
             return Collections.emptyList();
         }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
+        }
     }
 
     public static class ExceptionInstrumentation extends TracerAwareInstrumentation {
@@ -568,6 +578,11 @@ class InstrumentationTest {
         @Override
         public Collection<String> getInstrumentationGroupNames() {
             return Collections.emptyList();
+        }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
         }
     }
 
@@ -597,6 +612,11 @@ class InstrumentationTest {
         public Collection<String> getInstrumentationGroupNames() {
             return Collections.emptyList();
         }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
+        }
     }
 
     public static class FieldAccessInstrumentation extends TracerAwareInstrumentation {
@@ -620,6 +640,11 @@ class InstrumentationTest {
         @Override
         public Collection<String> getInstrumentationGroupNames() {
             return List.of("test", "experimental");
+        }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
         }
     }
 
@@ -645,6 +670,11 @@ class InstrumentationTest {
         public Collection<String> getInstrumentationGroupNames() {
             return List.of("test", "experimental");
         }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
+        }
     }
 
     public static class AssignToArgumentInstrumentation extends TracerAwareInstrumentation {
@@ -668,6 +698,11 @@ class InstrumentationTest {
         @Override
         public Collection<String> getInstrumentationGroupNames() {
             return List.of("test", "experimental");
+        }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
         }
     }
 
@@ -696,6 +731,11 @@ class InstrumentationTest {
         public Collection<String> getInstrumentationGroupNames() {
             return List.of("test", "experimental");
         }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
+        }
     }
 
     public static class AssignToReturnArrayInstrumentation extends TracerAwareInstrumentation {
@@ -719,6 +759,11 @@ class InstrumentationTest {
         @Override
         public Collection<String> getInstrumentationGroupNames() {
             return List.of("test", "experimental");
+        }
+
+        @Override
+        public boolean indyPlugin() {
+            return false;
         }
     }
 
@@ -1010,11 +1055,6 @@ class InstrumentationTest {
         @Override
         public Collection<String> getInstrumentationGroupNames() {
             return Collections.singletonList("test");
-        }
-
-        @Override
-        public boolean indyPlugin() {
-            return true;
         }
     }
