+++ apm-agent-plugins/apm-api-plugin/src/main/java/co/elastic/apm/agent/plugin/api/CaptureTransactionInstrumentation.java
@@ -120,4 +120,8 @@ public class CaptureTransactionInstrumentation extends TracerAwareInstrumentatio
         return Arrays.asList(PUBLIC_API_INSTRUMENTATION_GROUP, "annotations");
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
