+++ apm-agent-plugins/apm-api-plugin/src/main/java/co/elastic/apm/agent/plugin/api/TracedInstrumentation.java
@@ -132,4 +132,8 @@ public class TracedInstrumentation extends TracerAwareInstrumentation {
         return Arrays.asList(PUBLIC_API_INSTRUMENTATION_GROUP, "annotations");
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
