+++ apm-agent-plugins/apm-asynchttpclient-plugin/src/main/java/co/elastic/apm/agent/asynchttpclient/AbstractAsyncHttpClientInstrumentation.java
@@ -98,6 +98,11 @@ public abstract class AbstractAsyncHttpClientInstrumentation extends TracerAware
             .and(classLoaderCanLoadClass("org.asynchttpclient.AsyncHandler"));
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     public static class AsyncHttpClientInstrumentation extends AbstractAsyncHttpClientInstrumentation {
 
         public AsyncHttpClientInstrumentation(ElasticApmTracer tracer) {
