+++ apm-agent-plugins/apm-bootdelegation-plugin/src/main/java/co/elastic/apm/agent/bootdelegation/BootstrapDelegationClassLoaderInstrumentation.java
@@ -109,4 +109,9 @@ public class BootstrapDelegationClassLoaderInstrumentation extends TracerAwareIn
             }
         }
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
