+++ apm-agent-plugins/apm-dubbo-plugin/src/main/java/co/elastic/apm/agent/dubbo/AbstractDubboInstrumentation.java
@@ -35,4 +35,9 @@ public abstract class AbstractDubboInstrumentation extends TracerAwareInstrument
     public Collection<String> getInstrumentationGroupNames() {
         return Arrays.asList("dubbo", "experimental");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
