+++ apm-agent-plugins/apm-error-logging-plugin/src/main/java/co/elastic/apm/agent/error/logging/AbstractLoggerErrorCapturingInstrumentation.java
@@ -56,6 +56,11 @@ public abstract class AbstractLoggerErrorCapturingInstrumentation extends Tracer
         return LoggingAdvice.class;
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     public static class LoggingAdvice {
 
         @Advice.OnMethodEnter(suppress = Throwable.class)
