+++ apm-agent-plugins/apm-es-restclient-plugin/apm-es-restclient-plugin-common/src/main/java/co/elastic/apm/agent/es/restclient/ElasticsearchRestClientInstrumentation.java
@@ -59,4 +59,9 @@ public abstract class ElasticsearchRestClientInstrumentation extends TracerAware
     public Collection<String> getInstrumentationGroupNames() {
         return Collections.singleton("elasticsearch-restclient");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
