+++ apm-agent-plugins/apm-grails-plugin/src/main/java/co/elastic/apm/agent/grails/GrailsTransactionNameInstrumentation.java
@@ -85,6 +85,11 @@ public class GrailsTransactionNameInstrumentation extends TracerAwareInstrumenta
         return Collections.singletonList("grails");
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     @VisibleForAdvice
     public static class HandlerAdapterAdvice {
