+++ apm-agent-plugins/apm-grpc/apm-grpc-plugin/src/main/java/co/elastic/apm/agent/grpc/BaseInstrumentation.java
@@ -68,4 +68,8 @@ public abstract class BaseInstrumentation extends TracerAwareInstrumentation {
         return Collections.singleton(GrpcHelper.GRPC);
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
