+++ apm-agent-plugins/apm-hibernate-search-plugin/apm-hibernate-search-plugin-5_x/src/main/java/co/elastic/apm/agent/hibernate/search/v5_x/HibernateSearch5Instrumentation.java
@@ -83,6 +83,11 @@ public class HibernateSearch5Instrumentation extends TracerAwareInstrumentation 
         return Collections.singleton(HibernateSearchConstants.HIBERNATE_SEARCH_ORM_TYPE);
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     public static class HibernateSearch5ExecuteAdvice {
 
         @Advice.OnMethodEnter(suppress = Throwable.class)
