+++ apm-agent-plugins/apm-java-concurrent-plugin/src/main/java/co/elastic/apm/agent/concurrent/ExecutorInstrumentation.java
@@ -103,11 +103,6 @@ public abstract class ExecutorInstrumentation extends TracerAwareInstrumentation
         return Arrays.asList("concurrent", "executor");
     }
 
-    @Override
-    public boolean indyPlugin() {
-        return true;
-    }
-
     private static boolean isExcluded(@Advice.This Executor executor) {
         return excludedClasses.contains(executor.getClass().getName());
     }
