+++ apm-agent-plugins/apm-jaxws-plugin/src/main/java/co/elastic/apm/agent/jaxws/JaxWsTransactionNameInstrumentation.java
@@ -101,4 +101,9 @@ public class JaxWsTransactionNameInstrumentation extends TracerAwareInstrumentat
     public Collection<String> getInstrumentationGroupNames() {
         return Collections.singletonList("jax-ws");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
