+++ apm-agent-plugins/apm-kafka-plugin/apm-kafka-base-plugin/src/main/java/co/elastic/apm/agent/kafka/BaseKafkaInstrumentation.java
@@ -79,4 +79,9 @@ public abstract class BaseKafkaInstrumentation extends TracerAwareInstrumentatio
     public ElementMatcher.Junction<ClassLoader> getClassLoaderMatcher() {
         return not(isBootstrapClassLoader()).and(classLoaderCanLoadClass("org.apache.kafka.clients.consumer.ConsumerRecord"));
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
