+++ apm-agent-plugins/apm-mongoclient-plugin/src/main/java/co/elastic/apm/agent/mongoclient/MongoClientInstrumentation.java
@@ -48,4 +48,9 @@ public abstract class MongoClientInstrumentation extends TracerAwareInstrumentat
     public Collection<String> getInstrumentationGroupNames() {
         return Collections.singletonList("mongodb-client");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
