+++ apm-agent-plugins/apm-okhttp-plugin/src/main/java/co/elastic/apm/agent/okhttp/AbstractOkHttp3ClientInstrumentation.java
@@ -56,4 +56,9 @@ public abstract class AbstractOkHttp3ClientInstrumentation extends TracerAwareIn
     public Collection<String> getInstrumentationGroupNames() {
         return Arrays.asList("http-client", "okhttp");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
