+++ apm-agent-plugins/apm-opentracing-plugin/src/main/java/co/elastic/apm/agent/opentracing/impl/OpenTracingBridgeInstrumentation.java
@@ -39,4 +39,9 @@ public abstract class OpenTracingBridgeInstrumentation extends TracerAwareInstru
     public Collection<String> getInstrumentationGroupNames() {
         return Collections.singleton("opentracing");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
