+++ apm-agent-plugins/apm-process-plugin/src/main/java/co/elastic/apm/agent/process/BaseProcessInstrumentation.java
@@ -44,9 +44,4 @@ public abstract class BaseProcessInstrumentation extends TracerAwareInstrumentat
     public final Collection<String> getInstrumentationGroupNames() {
         return Collections.singletonList("process");
     }
-
-    @Override
-    public boolean indyPlugin() {
-        return true;
-    }
 }
