+++ apm-agent-plugins/apm-redis-plugin/apm-jedis-plugin/src/main/java/co/elastic/apm/agent/redis/jedis/JedisInstrumentation.java
@@ -89,4 +89,9 @@ public class JedisInstrumentation extends TracerAwareInstrumentation {
     public Collection<String> getInstrumentationGroupNames() {
         return Arrays.asList("redis", "jedis");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
