+++ apm-agent-plugins/apm-redis-plugin/apm-lettuce-plugin/src/main/java/co/elastic/apm/agent/redis/lettuce/Lettuce34Instrumentation.java
@@ -57,4 +57,9 @@ public abstract class Lettuce34Instrumentation extends TracerAwareInstrumentatio
     public Collection<String> getInstrumentationGroupNames() {
         return Arrays.asList("redis", "lettuce");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
