+++ apm-agent-plugins/apm-redis-plugin/apm-lettuce-plugin/src/main/java/co/elastic/apm/agent/redis/lettuce/Lettuce5StartSpanInstrumentation.java
@@ -92,4 +92,8 @@ public class Lettuce5StartSpanInstrumentation extends TracerAwareInstrumentation
         }
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
