+++ apm-agent-plugins/apm-redis-plugin/apm-redisson-plugin/src/main/java/co/elastic/apm/agent/redis/redisson/RedisConnectionInstrumentation.java
@@ -99,4 +99,9 @@ public class RedisConnectionInstrumentation extends TracerAwareInstrumentation {
     public Collection<String> getInstrumentationGroupNames() {
         return Arrays.asList("redis", "redisson");
     }
+
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
 }
