+++ apm-agent-plugins/apm-spring-webmvc-plugin/src/main/java/co/elastic/apm/agent/spring/webmvc/ExceptionHandlerInstrumentation.java
@@ -45,6 +45,11 @@ public class ExceptionHandlerInstrumentation extends TracerAwareInstrumentation 
         return ExceptionHandlerAdviceService.class;
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     public static class ExceptionHandlerAdviceService {
 
         @Advice.OnMethodEnter(suppress = Throwable.class)
