+++ apm-agent-plugins/apm-spring-webmvc-plugin/src/main/java/co/elastic/apm/agent/spring/webmvc/SpringServiceNameInstrumentation.java
@@ -67,6 +67,11 @@ public class SpringServiceNameInstrumentation extends TracerAwareInstrumentation
         return SpringServiceNameAdvice.class;
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     public static class SpringServiceNameAdvice {
 
         @Advice.OnMethodExit(suppress = Throwable.class)
