+++ apm-agent-plugins/apm-spring-webmvc-plugin/src/main/java/co/elastic/apm/agent/spring/webmvc/SpringTransactionNameInstrumentation.java
@@ -103,6 +103,11 @@ public class SpringTransactionNameInstrumentation extends TracerAwareInstrumenta
         return Collections.singleton("spring-mvc");
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     @VisibleForAdvice
     public static class HandlerAdapterAdvice {
