+++ apm-agent-plugins/apm-spring-webmvc-plugin/src/main/java/co/elastic/apm/agent/spring/webmvc/ViewRenderInstrumentation.java
@@ -58,6 +58,11 @@ public class ViewRenderInstrumentation extends TracerAwareInstrumentation {
         return ViewRenderAdviceService.class;
     }
 
+    @Override
+    public boolean indyPlugin() {
+        return false;
+    }
+
     public static class ViewRenderAdviceService {
 
         @Advice.OnMethodEnter(suppress = Throwable.class)
