+++ apm-agent-plugins/apm-urlconnection-plugin/src/main/java/co/elastic/apm/agent/urlconnection/SSLContextInstrumentation.java
@@ -77,11 +77,6 @@ public class SSLContextInstrumentation extends TracerAwareInstrumentation {
         return Collections.singleton("ssl-context");
     }
 
-    @Override
-    public boolean indyPlugin() {
-        return true;
-    }
-
     /**
      * This will not allow using the default SSL factory from any agent thread
      */
